jobs/requirements.txt
boto3==1.34.2
minio==7.2.0
python-dotenv==1.0.0
requests==2.31.0
